jenkins-build: Dynamically build helper containers
Instead of using the master tag of the helper containers, use those with
a tag that matches the repository revision.

Also build and deploy them if none are found.

Fixes balena-os/balena-yocto-scripts#179

Change-type: patch
Changelog-entry: Helper containers are now referenced by tag instead of using latest, and will be built and deployed if required
Signed-off-by: Alex Gonzalez <alexg@balena.io>